requirements.txt
discord.py
ibm_watson
